refactor: apply review suggestions and some more lints
