+++ src/token/cose/encrypted/encrypt/mod.rs
@@ -2,39 +2,42 @@ use alloc::rc::Rc;
 use alloc::vec::Vec;
 use core::cell::RefCell;
 
-use coset::{CoseMac, CoseMacBuilder, EncryptionContext, Header};
+use coset::{CoseEncrypt, CoseEncryptBuilder, EncryptionContext, Header};
 
 use crate::error::CoseCipherError;
-use crate::token::cose::encrypt::CoseKeyDistributionCipher;
+use crate::token::cose::encrypted;
+use crate::token::cose::encrypted::try_decrypt;
+use crate::token::cose::encrypted::{CoseEncryptCipher, CoseKeyDistributionCipher};
 use crate::token::cose::key::{CoseAadProvider, CoseKeyProvider};
-use crate::token::cose::mac::{try_compute, try_verify, CoseMacCipher};
-use crate::token::cose::recipient::CoseNestedRecipientSearchContext;
+use crate::token::cose::recipient::{
+    struct_to_recipient_context, CoseNestedRecipientSearchContext,
+};
 
 #[cfg(all(test, feature = "std"))]
 mod tests;
 
-pub trait CoseMacBuilderExt: Sized {
-    fn try_compute<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+pub trait CoseEncryptBuilderExt: Sized {
+    fn try_encrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         protected: Option<Header>,
         unprotected: Option<Header>,
-        payload: Vec<u8>,
+        plaintext: &[u8],
         external_aad: &mut CAP,
     ) -> Result<Self, CoseCipherError<B::Error>>;
 }
 
-impl CoseMacBuilderExt for CoseMacBuilder {
-    fn try_compute<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+impl CoseEncryptBuilderExt for CoseEncryptBuilder {
+    fn try_encrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         protected: Option<Header>,
         unprotected: Option<Header>,
-        payload: Vec<u8>,
+        plaintext: &[u8],
         external_aad: &mut CAP,
     ) -> Result<Self, CoseCipherError<B::Error>> {
         let mut builder = self;
@@ -44,36 +47,39 @@ impl CoseMacBuilderExt for CoseMacBuilder {
         if let Some(unprotected) = &unprotected {
             builder = builder.unprotected(unprotected.clone());
         }
-        builder = builder.payload(payload);
-        Ok(builder.create_tag(
-            external_aad.lookup_aad(None, protected.as_ref(), unprotected.as_ref()),
-            |input| {
-                // TODO proper error handling here
-                try_compute(
+        builder.try_create_ciphertext(
+            plaintext,
+            external_aad.lookup_aad(
+                Some(EncryptionContext::CoseEncrypt),
+                protected.as_ref(),
+                unprotected.as_ref(),
+            ),
+            |plaintext, aad| {
+                encrypted::try_encrypt(
                     backend,
                     key_provider,
                     protected.as_ref(),
                     unprotected.as_ref(),
                     try_all_keys,
-                    input,
+                    plaintext,
+                    aad,
                 )
-                .expect("computing MAC failed")
             },
-        ))
+        )
     }
 }
 
-pub trait CoseMacExt {
-    fn try_verify<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+pub trait CoseEncryptExt {
+    fn try_decrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         &self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<(), CoseCipherError<B::Error>>;
+    ) -> Result<Vec<u8>, CoseCipherError<B::Error>>;
 
-    fn try_verify_with_recipients<
-        B: CoseKeyDistributionCipher + CoseMacCipher,
+    fn try_decrypt_with_recipients<
+        B: CoseKeyDistributionCipher + CoseEncryptCipher,
         CKP: CoseKeyProvider,
         CAP: CoseAadProvider,
     >(
@@ -82,37 +88,41 @@ pub trait CoseMacExt {
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<(), CoseCipherError<B::Error>>;
+    ) -> Result<Vec<u8>, CoseCipherError<B::Error>>;
 }
 
-impl CoseMacExt for CoseMac {
-    fn try_verify<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+impl CoseEncryptExt for CoseEncrypt {
+    fn try_decrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         &self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<(), CoseCipherError<B::Error>> {
+    ) -> Result<Vec<u8>, CoseCipherError<B::Error>> {
         let backend = Rc::new(RefCell::new(backend));
         let key_provider = Rc::new(RefCell::new(key_provider));
-        self.verify_tag(
-            external_aad.lookup_aad(None, Some(&self.protected.header), Some(&self.unprotected)),
-            |tag, input| {
-                try_verify(
+        self.decrypt(
+            external_aad.lookup_aad(
+                Some(EncryptionContext::CoseEncrypt),
+                Some(&self.protected.header),
+                Some(&self.unprotected),
+            ),
+            |ciphertext, aad| {
+                encrypted::try_decrypt(
                     &backend,
                     &key_provider,
                     &self.protected.header,
                     &self.unprotected,
                     try_all_keys,
-                    tag,
-                    input,
+                    ciphertext,
+                    aad,
                 )
             },
         )
     }
 
-    fn try_verify_with_recipients<
-        B: CoseKeyDistributionCipher + CoseMacCipher,
+    fn try_decrypt_with_recipients<
+        B: CoseKeyDistributionCipher + CoseEncryptCipher,
         CKP: CoseKeyProvider,
         CAP: CoseAadProvider,
     >(
@@ -121,7 +131,7 @@ impl CoseMacExt for CoseMac {
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<(), CoseCipherError<B::Error>> {
+    ) -> Result<Vec<u8>, CoseCipherError<B::Error>> {
         let backend = Rc::new(RefCell::new(backend));
         let key_provider = Rc::new(RefCell::new(key_provider));
         let mut nested_recipient_key_provider = CoseNestedRecipientSearchContext::new(
@@ -129,19 +139,23 @@ impl CoseMacExt for CoseMac {
             Rc::clone(&backend),
             Rc::clone(&key_provider),
             try_all_keys,
-            EncryptionContext::MacRecipient,
+            struct_to_recipient_context(EncryptionContext::CoseEncrypt),
         );
-        self.verify_tag(
-            external_aad.lookup_aad(None, Some(&self.protected.header), Some(&self.unprotected)),
-            |tag, input| {
-                try_verify(
+        self.decrypt(
+            external_aad.lookup_aad(
+                Some(EncryptionContext::CoseEncrypt),
+                Some(&self.protected.header),
+                Some(&self.unprotected),
+            ),
+            |ciphertext, aad| {
+                try_decrypt(
                     &backend,
                     &Rc::new(RefCell::new(&mut nested_recipient_key_provider)),
                     &self.protected.header,
                     &self.unprotected,
                     true,
-                    tag,
-                    input,
+                    ciphertext,
+                    aad,
                 )
             },
         )
