+++ src/token/cose/maced/mac/mod.rs
@@ -2,42 +2,39 @@ use alloc::rc::Rc;
 use alloc::vec::Vec;
 use core::cell::RefCell;
 
-use coset::{CoseEncrypt, CoseEncryptBuilder, EncryptionContext, Header};
+use coset::{CoseMac, CoseMacBuilder, EncryptionContext, Header};
 
 use crate::error::CoseCipherError;
-use crate::token::cose::encrypt;
-use crate::token::cose::encrypt::try_decrypt;
-use crate::token::cose::encrypt::{CoseEncryptCipher, CoseKeyDistributionCipher};
+use crate::token::cose::encrypted::CoseKeyDistributionCipher;
 use crate::token::cose::key::{CoseAadProvider, CoseKeyProvider};
-use crate::token::cose::recipient::{
-    struct_to_recipient_context, CoseNestedRecipientSearchContext,
-};
+use crate::token::cose::maced::{try_compute, try_verify, CoseMacCipher};
+use crate::token::cose::recipient::CoseNestedRecipientSearchContext;
 
 #[cfg(all(test, feature = "std"))]
 mod tests;
 
-pub trait CoseEncryptBuilderExt: Sized {
-    fn try_encrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+pub trait CoseMacBuilderExt: Sized {
+    fn try_compute<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         protected: Option<Header>,
         unprotected: Option<Header>,
-        plaintext: &[u8],
+        payload: Vec<u8>,
         external_aad: &mut CAP,
     ) -> Result<Self, CoseCipherError<B::Error>>;
 }
 
-impl CoseEncryptBuilderExt for CoseEncryptBuilder {
-    fn try_encrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+impl CoseMacBuilderExt for CoseMacBuilder {
+    fn try_compute<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         protected: Option<Header>,
         unprotected: Option<Header>,
-        plaintext: &[u8],
+        payload: Vec<u8>,
         external_aad: &mut CAP,
     ) -> Result<Self, CoseCipherError<B::Error>> {
         let mut builder = self;
@@ -47,39 +44,36 @@ impl CoseEncryptBuilderExt for CoseEncryptBuilder {
         if let Some(unprotected) = &unprotected {
             builder = builder.unprotected(unprotected.clone());
         }
-        builder.try_create_ciphertext(
-            plaintext,
-            external_aad.lookup_aad(
-                Some(EncryptionContext::CoseEncrypt),
-                protected.as_ref(),
-                unprotected.as_ref(),
-            ),
-            |plaintext, aad| {
-                encrypt::try_encrypt(
+        builder = builder.payload(payload);
+        Ok(builder.create_tag(
+            external_aad.lookup_aad(None, protected.as_ref(), unprotected.as_ref()),
+            |input| {
+                // TODO proper error handling here
+                try_compute(
                     backend,
                     key_provider,
                     protected.as_ref(),
                     unprotected.as_ref(),
                     try_all_keys,
-                    plaintext,
-                    aad,
+                    input,
                 )
+                .expect("computing MAC failed")
             },
-        )
+        ))
     }
 }
 
-pub trait CoseEncryptExt {
-    fn try_decrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+pub trait CoseMacExt {
+    fn try_verify<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         &self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<Vec<u8>, CoseCipherError<B::Error>>;
+    ) -> Result<(), CoseCipherError<B::Error>>;
 
-    fn try_decrypt_with_recipients<
-        B: CoseKeyDistributionCipher + CoseEncryptCipher,
+    fn try_verify_with_recipients<
+        B: CoseKeyDistributionCipher + CoseMacCipher,
         CKP: CoseKeyProvider,
         CAP: CoseAadProvider,
     >(
@@ -88,41 +82,37 @@ pub trait CoseEncryptExt {
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<Vec<u8>, CoseCipherError<B::Error>>;
+    ) -> Result<(), CoseCipherError<B::Error>>;
 }
 
-impl CoseEncryptExt for CoseEncrypt {
-    fn try_decrypt<B: CoseEncryptCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
+impl CoseMacExt for CoseMac {
+    fn try_verify<B: CoseMacCipher, CKP: CoseKeyProvider, CAP: CoseAadProvider>(
         &self,
         backend: &mut B,
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<Vec<u8>, CoseCipherError<B::Error>> {
+    ) -> Result<(), CoseCipherError<B::Error>> {
         let backend = Rc::new(RefCell::new(backend));
         let key_provider = Rc::new(RefCell::new(key_provider));
-        self.decrypt(
-            external_aad.lookup_aad(
-                Some(EncryptionContext::CoseEncrypt),
-                Some(&self.protected.header),
-                Some(&self.unprotected),
-            ),
-            |ciphertext, aad| {
-                encrypt::try_decrypt(
+        self.verify_tag(
+            external_aad.lookup_aad(None, Some(&self.protected.header), Some(&self.unprotected)),
+            |tag, input| {
+                try_verify(
                     &backend,
                     &key_provider,
                     &self.protected.header,
                     &self.unprotected,
                     try_all_keys,
-                    ciphertext,
-                    aad,
+                    tag,
+                    input,
                 )
             },
         )
     }
 
-    fn try_decrypt_with_recipients<
-        B: CoseKeyDistributionCipher + CoseEncryptCipher,
+    fn try_verify_with_recipients<
+        B: CoseKeyDistributionCipher + CoseMacCipher,
         CKP: CoseKeyProvider,
         CAP: CoseAadProvider,
     >(
@@ -131,7 +121,7 @@ impl CoseEncryptExt for CoseEncrypt {
         key_provider: &mut CKP,
         try_all_keys: bool,
         external_aad: &mut CAP,
-    ) -> Result<Vec<u8>, CoseCipherError<B::Error>> {
+    ) -> Result<(), CoseCipherError<B::Error>> {
         let backend = Rc::new(RefCell::new(backend));
         let key_provider = Rc::new(RefCell::new(key_provider));
         let mut nested_recipient_key_provider = CoseNestedRecipientSearchContext::new(
@@ -139,23 +129,19 @@ impl CoseEncryptExt for CoseEncrypt {
             Rc::clone(&backend),
             Rc::clone(&key_provider),
             try_all_keys,
-            struct_to_recipient_context(EncryptionContext::CoseEncrypt),
+            EncryptionContext::MacRecipient,
         );
-        self.decrypt(
-            external_aad.lookup_aad(
-                Some(EncryptionContext::CoseEncrypt),
-                Some(&self.protected.header),
-                Some(&self.unprotected),
-            ),
-            |ciphertext, aad| {
-                try_decrypt(
+        self.verify_tag(
+            external_aad.lookup_aad(None, Some(&self.protected.header), Some(&self.unprotected)),
+            |tag, input| {
+                try_verify(
                     &backend,
                     &Rc::new(RefCell::new(&mut nested_recipient_key_provider)),
                     &self.protected.header,
                     &self.unprotected,
                     true,
-                    ciphertext,
-                    aad,
+                    tag,
+                    input,
                 )
             },
         )
