+++ src/token/cose/maced/mac/tests.rs
@@ -3,79 +3,97 @@ use core::convert::Infallible;
 use std::path::PathBuf;
 
 use coset::iana::Algorithm;
-use coset::{CoseError, CoseKey, CoseKeyBuilder, CoseMac0, CoseMac0Builder, Header};
+use coset::{
+    CoseError, CoseKey, CoseKeyBuilder, CoseMac, CoseMacBuilder, CoseRecipientBuilder,
+    EncryptionContext, Header,
+};
 use rstest::rstest;
 
 use crate::token::cose::crypto_impl::openssl::OpensslContext;
+use crate::token::cose::encrypted::CoseKeyDistributionCipher;
 use crate::token::cose::header_util::determine_algorithm;
-use crate::token::cose::mac::mac0::{CoseMac0BuilderExt, CoseMac0Ext};
-use crate::token::cose::mac::CoseMacCipher;
+use crate::token::cose::key::CoseSymmetricKey;
+use crate::token::cose::maced::mac::{CoseMacBuilderExt, CoseMacExt};
+use crate::token::cose::maced::CoseMacCipher;
+use crate::token::cose::recipient::CoseRecipientBuilderExt;
 use crate::token::cose::test_helper::{
     apply_attribute_failures, apply_header_failures, serialize_cose_with_failures,
     CoseStructTestHelper, TestCase,
 };
 use crate::token::cose::{test_helper, CoseCipher};
 
-impl<B: CoseCipher + CoseMacCipher> CoseStructTestHelper<B> for CoseMac0 {
+impl<B: CoseCipher + CoseMacCipher + CoseKeyDistributionCipher> CoseStructTestHelper<B>
+    for CoseMac
+{
     fn from_test_case(case: &TestCase, backend: &mut B) -> Self {
-        let mac0_cfg = case
+        let mac_cfg = case
             .input
-            .mac0
+            .mac
             .as_ref()
-            .expect("expected a CoseMac0 test case, but it was not found");
+            .expect("expected a CoseEncrypt test case, but it was not found");
 
-        let mac0 = CoseMac0Builder::new();
+        let mac = CoseMacBuilder::new();
 
-        let recipient = mac0_cfg
+        let recipient = mac_cfg
             .recipients
             .first()
             .expect("test case has no recipient");
 
-        let unprotected = mac0_cfg.unprotected.clone().unwrap_or_default();
+        let unprotected = mac_cfg.unprotected.clone().unwrap_or_default();
 
-        let enc_key = if recipient.alg == Some(Algorithm::Direct)
+        let mut recipient_struct_builder = CoseRecipientBuilder::from(recipient.clone());
+        let enc_key: CoseKey;
+        if recipient.alg == Some(Algorithm::Direct)
             || determine_algorithm::<Infallible>(
                 None,
                 recipient.unprotected.as_ref(),
                 recipient.protected.as_ref(),
             ) == Ok(coset::Algorithm::Assigned(Algorithm::Direct))
         {
-            recipient.key.clone()
+            enc_key = recipient.key.clone();
         } else {
-            CoseKeyBuilder::new_symmetric_key(
+            enc_key = CoseKeyBuilder::new_symmetric_key(
                 case.intermediates
                     .as_ref()
-                    .expect("CoseMac0 test case should have intermediates")
+                    .expect("CoseEncrypt test case should have intermediates")
                     .cek
                     .clone(),
             )
-            .build()
-        };
+            .build();
+            let parsed_key = CoseSymmetricKey::<Infallible>::try_from(&enc_key)
+                .expect("unable to parse CEK input as symmetric key");
+            recipient_struct_builder = recipient_struct_builder
+                .try_encrypt(
+                    backend,
+                    &mut &recipient.key,
+                    true,
+                    EncryptionContext::EncRecipient,
+                    recipient.protected.clone(),
+                    recipient.unprotected.clone(),
+                    parsed_key.k,
+                    &mut (&[] as &[u8]),
+                )
+                .expect("unable to create CoseRecipient structure");
+        }
 
-        let mac0 = mac0
+        mac.add_recipient(recipient_struct_builder.build())
             .try_compute(
                 backend,
                 &mut &enc_key,
                 false,
-                mac0_cfg.protected.clone(),
+                mac_cfg.protected.clone(),
                 Some(unprotected),
                 case.input.plaintext.clone().into_bytes(),
-                &mut mac0_cfg.external.as_slice(),
+                &mut mac_cfg.external.as_slice(),
             )
-            .expect("unable to encrypt Mac0 object");
-
-        mac0.build()
+            .expect("unable to encrypt Encrypt object")
+            .build()
     }
 
     fn serialize_and_apply_failures(mut self, case: &TestCase) -> Result<Vec<u8>, CoseError> {
         let failures = &case.input.failures;
         if let Some(1) = &failures.change_tag {
-            let byte = self
-                .payload
-                .as_mut()
-                .expect("Mac0 has no payload, can't apply failure")
-                .first_mut()
-                .unwrap();
+            let byte = self.tag.first_mut().unwrap();
             *byte = byte.wrapping_add(1);
         }
 
@@ -86,11 +104,7 @@ impl<B: CoseCipher + CoseMacCipher> CoseStructTestHelper<B> for CoseMac0 {
     }
 
     fn check_against_test_case(&self, case: &TestCase, backend: &mut B) {
-        let test_case = case
-            .input
-            .mac0
-            .as_ref()
-            .expect("CoseMac0 test case expected");
+        let test_case = case.input.mac.as_ref().expect("CoseMac test case expected");
         let keys: Vec<CoseKey> = test_case
             .recipients
             .iter()
@@ -104,7 +118,7 @@ impl<B: CoseCipher + CoseMacCipher> CoseStructTestHelper<B> for CoseMac0 {
             .collect();
         let mut aad = test_case.external.as_slice();
 
-        let verify_result = self.try_verify(backend, &mut &keys, false, &mut aad);
+        let verify_result = self.try_verify_with_recipients(backend, &mut &keys, false, &mut aad);
 
         if case.fail {
             verify_result.expect_err("invalid token was successfully verified");
@@ -129,17 +143,17 @@ impl<B: CoseCipher + CoseMacCipher> CoseStructTestHelper<B> for CoseMac0 {
 }
 
 #[rstest]
-fn cose_examples_mac0_reference_output<B: CoseMacCipher>(
-    #[files("tests/cose_examples/mac0-tests/mac-*.json")] test_path: PathBuf,
+fn cose_examples_mac_reference_output<B: CoseMacCipher + CoseKeyDistributionCipher>(
+    #[files("tests/cose_examples/mac-tests/mac-*.json")] test_path: PathBuf,
     #[values(OpensslContext {})] backend: B,
 ) {
-    test_helper::perform_cose_reference_output_test::<CoseMac0, B>(test_path, backend);
+    test_helper::perform_cose_reference_output_test::<CoseMac, B>(test_path, backend);
 }
 
 #[rstest]
-fn cose_examples_mac0_self_signed<B: CoseMacCipher>(
-    #[files("tests/cose_examples/mac0-tests/mac-*.json")] test_path: PathBuf,
+fn cose_examples_mac_self_signed<B: CoseMacCipher + CoseKeyDistributionCipher>(
+    #[files("tests/cose_examples/mac-tests/mac-*.json")] test_path: PathBuf,
     #[values(OpensslContext {})] backend: B,
 ) {
-    test_helper::perform_cose_self_signed_test::<CoseMac0, B>(test_path, backend);
+    test_helper::perform_cose_self_signed_test::<CoseMac, B>(test_path, backend);
 }
