+++ src/token/cose/maced/mac0/tests.rs
@@ -3,97 +3,79 @@ use core::convert::Infallible;
 use std::path::PathBuf;
 
 use coset::iana::Algorithm;
-use coset::{
-    CoseError, CoseKey, CoseKeyBuilder, CoseMac, CoseMacBuilder, CoseRecipientBuilder,
-    EncryptionContext, Header,
-};
+use coset::{CoseError, CoseKey, CoseKeyBuilder, CoseMac0, CoseMac0Builder, Header};
 use rstest::rstest;
 
 use crate::token::cose::crypto_impl::openssl::OpensslContext;
-use crate::token::cose::encrypt::CoseKeyDistributionCipher;
 use crate::token::cose::header_util::determine_algorithm;
-use crate::token::cose::key::CoseSymmetricKey;
-use crate::token::cose::mac::mac::{CoseMacBuilderExt, CoseMacExt};
-use crate::token::cose::mac::CoseMacCipher;
-use crate::token::cose::recipient::CoseRecipientBuilderExt;
+use crate::token::cose::maced::mac0::{CoseMac0BuilderExt, CoseMac0Ext};
+use crate::token::cose::maced::CoseMacCipher;
 use crate::token::cose::test_helper::{
     apply_attribute_failures, apply_header_failures, serialize_cose_with_failures,
     CoseStructTestHelper, TestCase,
 };
 use crate::token::cose::{test_helper, CoseCipher};
 
-impl<B: CoseCipher + CoseMacCipher + CoseKeyDistributionCipher> CoseStructTestHelper<B>
-    for CoseMac
-{
+impl<B: CoseCipher + CoseMacCipher> CoseStructTestHelper<B> for CoseMac0 {
     fn from_test_case(case: &TestCase, backend: &mut B) -> Self {
-        let mac_cfg = case
+        let mac0_cfg = case
             .input
-            .mac
+            .mac0
             .as_ref()
-            .expect("expected a CoseEncrypt test case, but it was not found");
+            .expect("expected a CoseMac0 test case, but it was not found");
 
-        let mac = CoseMacBuilder::new();
+        let mac0 = CoseMac0Builder::new();
 
-        let recipient = mac_cfg
+        let recipient = mac0_cfg
             .recipients
             .first()
             .expect("test case has no recipient");
 
-        let unprotected = mac_cfg.unprotected.clone().unwrap_or_default();
+        let unprotected = mac0_cfg.unprotected.clone().unwrap_or_default();
 
-        let mut recipient_struct_builder = CoseRecipientBuilder::from(recipient.clone());
-        let enc_key: CoseKey;
-        if recipient.alg == Some(Algorithm::Direct)
+        let enc_key = if recipient.alg == Some(Algorithm::Direct)
             || determine_algorithm::<Infallible>(
                 None,
                 recipient.unprotected.as_ref(),
                 recipient.protected.as_ref(),
             ) == Ok(coset::Algorithm::Assigned(Algorithm::Direct))
         {
-            enc_key = recipient.key.clone();
+            recipient.key.clone()
         } else {
-            enc_key = CoseKeyBuilder::new_symmetric_key(
+            CoseKeyBuilder::new_symmetric_key(
                 case.intermediates
                     .as_ref()
-                    .expect("CoseEncrypt test case should have intermediates")
+                    .expect("CoseMac0 test case should have intermediates")
                     .cek
                     .clone(),
             )
-            .build();
-            let parsed_key = CoseSymmetricKey::<Infallible>::try_from(&enc_key)
-                .expect("unable to parse CEK input as symmetric key");
-            recipient_struct_builder = recipient_struct_builder
-                .try_encrypt(
-                    backend,
-                    &mut &recipient.key,
-                    true,
-                    EncryptionContext::EncRecipient,
-                    recipient.protected.clone(),
-                    recipient.unprotected.clone(),
-                    parsed_key.k,
-                    &mut (&[] as &[u8]),
-                )
-                .expect("unable to create CoseRecipient structure");
-        }
+            .build()
+        };
 
-        mac.add_recipient(recipient_struct_builder.build())
+        let mac0 = mac0
             .try_compute(
                 backend,
                 &mut &enc_key,
                 false,
-                mac_cfg.protected.clone(),
+                mac0_cfg.protected.clone(),
                 Some(unprotected),
                 case.input.plaintext.clone().into_bytes(),
-                &mut mac_cfg.external.as_slice(),
+                &mut mac0_cfg.external.as_slice(),
             )
-            .expect("unable to encrypt Encrypt object")
-            .build()
+            .expect("unable to encrypt Mac0 object");
+
+        mac0.build()
     }
 
     fn serialize_and_apply_failures(mut self, case: &TestCase) -> Result<Vec<u8>, CoseError> {
         let failures = &case.input.failures;
         if let Some(1) = &failures.change_tag {
-            let byte = self.tag.first_mut().unwrap();
+            let byte = self
+                .payload
+                .as_mut()
+                .expect("Mac0 has no payload, can't apply failure")
+                .first_mut()
+                .unwrap();
             *byte = byte.wrapping_add(1);
         }
 
@@ -104,7 +86,11 @@ impl<B: CoseCipher + CoseMacCipher + CoseKeyDistributionCipher> CoseStructTestHe
     }
 
     fn check_against_test_case(&self, case: &TestCase, backend: &mut B) {
-        let test_case = case.input.mac.as_ref().expect("CoseMac test case expected");
+        let test_case = case
+            .input
+            .mac0
+            .as_ref()
+            .expect("CoseMac0 test case expected");
         let keys: Vec<CoseKey> = test_case
             .recipients
             .iter()
@@ -118,7 +104,7 @@ impl<B: CoseCipher + CoseMacCipher + CoseKeyDistributionCipher> CoseStructTestHe
             .collect();
         let mut aad = test_case.external.as_slice();
 
-        let verify_result = self.try_verify_with_recipients(backend, &mut &keys, false, &mut aad);
+        let verify_result = self.try_verify(backend, &mut &keys, false, &mut aad);
 
         if case.fail {
             verify_result.expect_err("invalid token was successfully verified");
@@ -143,17 +129,17 @@ impl<B: CoseCipher + CoseMacCipher + CoseKeyDistributionCipher> CoseStructTestHe
 }
 
 #[rstest]
-fn cose_examples_mac_reference_output<B: CoseMacCipher + CoseKeyDistributionCipher>(
-    #[files("tests/cose_examples/mac-tests/mac-*.json")] test_path: PathBuf,
+fn cose_examples_mac0_reference_output<B: CoseMacCipher>(
+    #[files("tests/cose_examples/mac0-tests/mac-*.json")] test_path: PathBuf,
     #[values(OpensslContext {})] backend: B,
 ) {
-    test_helper::perform_cose_reference_output_test::<CoseMac, B>(test_path, backend);
+    test_helper::perform_cose_reference_output_test::<CoseMac0, B>(test_path, backend);
 }
 
 #[rstest]
-fn cose_examples_mac_self_signed<B: CoseMacCipher + CoseKeyDistributionCipher>(
-    #[files("tests/cose_examples/mac-tests/mac-*.json")] test_path: PathBuf,
+fn cose_examples_mac0_self_signed<B: CoseMacCipher>(
+    #[files("tests/cose_examples/mac0-tests/mac-*.json")] test_path: PathBuf,
     #[values(OpensslContext {})] backend: B,
 ) {
-    test_helper::perform_cose_self_signed_test::<CoseMac, B>(test_path, backend);
+    test_helper::perform_cose_self_signed_test::<CoseMac0, B>(test_path, backend);
 }
