+++ src/token/cose/signed/sign/tests.rs
@@ -6,8 +6,8 @@ use coset::{CoseError, CoseKey, CoseSign, CoseSignBuilder, CoseSignatureBuilder,
 use rstest::rstest;
 
 use crate::token::cose::crypto_impl::openssl::OpensslContext;
-use crate::token::cose::encrypt::CoseKeyDistributionCipher;
-use crate::token::cose::sign::{CoseSignBuilderExt, CoseSignExt};
+use crate::token::cose::encrypted::CoseKeyDistributionCipher;
+use crate::token::cose::signed::{CoseSignBuilderExt, CoseSignExt};
 use crate::token::cose::test_helper::{
     apply_attribute_failures, apply_header_failures, perform_cose_reference_output_test,
     perform_cose_self_signed_test, serialize_cose_with_failures, CoseStructTestHelper, TestCase,
